[CL-1660] Fix tooltips language and type cards
