+++ front/app/modules/commercial/analytics/admin/components/VisitorsTypeCard/messages.ts
@@ -5,4 +5,8 @@ export default defineMessages({
     id: 'app.modules.commercial.analytics.admin.components.VisitorsTypesCard.title',
     defaultMessage: 'Visitor Types',
   },
+  visitors: {
+    id: 'app.modules.commercial.analytics.admin.components.VisitorsTypesCard.visitors',
+    defaultMessage: 'Visitors',
+  },
 });
